requirements.txt
torch
transformers
sentencepiece
pandas
python-docx
PyPDF2
openpyxl
pytesseract
Pillow
streamlit
